+++ packages/nocodb-sdk/src/lib/sqlUi/SqliteUi.ts
@@ -850,7 +850,7 @@ export class SqliteUi {
         };
         break;
       case 'Percent':
-        colProp.dt = 'double';
+        colProp.dt = 'decimal';
         break;
       case 'Duration':
         colProp.dt = 'decimal';
